@@ -15,22 +15,21 @@
 	},
 	"dependencies": {
 		"@crowdin/cli": "3",
-		"@docusaurus/core": "2.4.3",
-		"@docusaurus/plugin-client-redirects": "2.4.3",
-		"@docusaurus/plugin-content-docs": "2.4.3",
-		"@docusaurus/plugin-pwa": "2.4.3",
-		"@docusaurus/preset-classic": "2.4.3",
+		"@docusaurus/core": "3.2.1",
+		"@docusaurus/plugin-client-redirects": "3.2.1",
+		"@docusaurus/plugin-content-docs": "3.2.1",
+		"@docusaurus/plugin-pwa": "3.2.1",
+		"@docusaurus/preset-classic": "3.2.1",
 		"@easyops-cn/docusaurus-search-local": "*",
-		"@mdx-js/react": "^1.6",
-		"@saucelabs/theme-github-codeblock": "^0",
-		"clsx": "^1.2",
+		"@mdx-js/react": "^3",
+		"clsx": "^2",
 		"docusaurus-plugin-image-zoom": "^1",
 		"docusaurus-plugin-matomo": "^0",
-		"prism-react-renderer": "^1.3",
-		"react": "^17",
-		"react-dom": "^17",
+		"prism-react-renderer": "^2",
+		"react": "^18",
+		"react-dom": "^18",
 		"react-player": "^2",
-		"redocusaurus": "^1",
+		"redocusaurus": "^2",
 		"url": "^0"
 	}
 }
